services/gateway/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
pydantic==2.*
httpx==0.27.*
pytest==8.*
pytest-asyncio==0.23.*
pydantic-settings==2.*
PyJWT
redis==5.*
SQLAlchemy>=2.0
asyncpg>=0.29 
pgvector>=0.2
nats-py
prometheus-client
bcrypt>=4.0.1
websockets>=11.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

